mainwindow: add in UI
If the login is successful, an information message is displayed and the username is displayed in the status bar. Depending on the user's role, certain elements of the user interface are hidden or shown.

The data for tanks is retrieved from the database and the interface is filled with it.

The price of the fuel is displayed in a lineEdit (and an admin can modify it)